# src/config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/api/v1.0",
  "schemes": ["http"],
  "produces": ["application/json"],
  "paths": {
    "/contacts": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Retrieves all contacts.",
        "responses": {
          "200": { "description": "OK", "schema": { "$ref": "#/definitions/Contacts" } },
          "400": { "description": "Bad request", "schema": { "$ref": "#/definitions/Error" } },
          "500": { "description": "Internal server error", "schema": { "$ref": "#/definitions/Error" } }
        },
        "content": { "application/json": {} },
        "parameters": []
      },
      "post": {
        "tags": ["Contacts"],
        "summary": "Creates contact.",
        "responses": {
          "200": { "description": "OK", "schema": { "$ref": "#/definitions/Contact" } },
          "400": { "description": "Bad request", "schema": { "$ref": "#/definitions/Error" } },
          "500": { "description": "Internal server error", "schema": { "$ref": "#/definitions/Error" } }
        },
        "content": { "application/json": {} },
        "parameters": [
          {
            "name": "contact",
            "in": "body",
            "required": true,
            "description": "Contact to add to the store.",
            "schema": { "$ref": "#/definitions/Contact" }
          }
        ]
      }
    },
    "/contacts/{contactID}": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Retrieves contact.",
        "responses": {
          "200": { "description": "OK", "schema": { "$ref": "#/definitions/Contact" } },
          "400": { "description": "Bad request", "schema": { "$ref": "#/definitions/Error" } },
          "500": { "description": "Internal server error", "schema": { "$ref": "#/definitions/Error" } }
        },
        "content": { "application/json": {} },
        "parameters": [
          {
            "name": "contactID",
            "in": "path",
            "required": true,
            "description": "ID of contact to fetch.",
            "type": "string"
          }
        ]
      },
      "patch": {
        "tags": ["Contacts"],
        "summary": "Updates contact.",
        "responses": {
          "200": { "description": "OK", "schema": { "$ref": "#/definitions/Contact" } },
          "400": { "description": "Bad request", "schema": { "$ref": "#/definitions/Error" } },
          "500": { "description": "Internal server error", "schema": { "$ref": "#/definitions/Error" } }
        },
        "content": { "application/json": {} },
        "parameters": [
          {
            "name": "contactID",
            "in": "path",
            "required": true,
            "description": "ID of contact to update.",
            "type": "string"
          },
          {
            "name": "changes",
            "in": "body",
            "required": true,
            "description": "Changes to update the contact with.",
            "schema": { "$ref": "#/definitions/ContactUpdate" }
          }
        ]
      }
    }
  },
  "definitions": {
    "Contact": {
      "type": "object",
      "required": ["id", "username"],
      "properties": {
        "id": { "type": "string" },
        "username": { "type": "string" },
        "firstName": { "type": "string" },
        "lastName": { "type": "string" },
        "email": { "type": "string" },
        "address": { "type": "string" },
        "county": { "type": "string" },
        "country": { "type": "string" },
        "postcode": { "type": "string" }
      }
    },
    "ContactUpdate": {
      "type": "object",
      "properties": {
        "username": { "type": "string" },
        "firstName": { "type": "string" },
        "lastName": { "type": "string" },
        "email": { "type": "string" },
        "address": { "type": "string" },
        "county": { "type": "string" },
        "country": { "type": "string" },
        "postcode": { "type": "string" }
      }
    },
    "Contacts": {
      "type": "array",
      "items": { "$ref": "#/definitions/Contact" }
    },
    "Error": {
      "type": "object",
      "required": ["code", "message"],
      "properties": { "code": { "type": "integer", "format": "int32" }, "message": { "type": "string" } }
    }
  }
}